add_single_live_event доработки пра
