+++ mvi-core/src/main/kotlin/ru/surfstudio/mvi/vm/compose/FoundationMviCompose.kt
@@ -24,6 +24,9 @@ fun interface ComposedViewContext<E : Event> {
     fun emit(event: E)
 }
 
-interface CommandEmmiter<C : CommandEvent> {
-    val emitCommandCallback: (C) -> Unit
+/**
+ *  Helpful interface for emitting command event from [Reducer]
+ */
+ interface CommandEmmiter<E> where E : Event, E : CommandEvent  {
+    val emitCommandCallback: (E) -> Unit
 }
